Restructuring the files to accomodate for adding replies
