+++ server/forum/thread/getThread/getThreads.repository.js
@@ -1,4 +1,4 @@
-const pool = require("../db");
+const pool = require("../../../db");
 
 /* 
 -----------------------------------
